+++ core/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Create & tear down a new Docker network per test, to pave the way for parallel tests
 * Move FreeIpAddrTracker a little closer to handling IPv6
 * Run tests in parallel!
+* Print errors directly, rather than rendering them through logrus, to preserve newlines
 
 # 0.4.0
 * remove freeHostPortTracker and all host-container port mappings